% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTools.R
\name{f_tbyl}
\alias{f_tbyl}
\title{Provides a janitor distribution table but with formatting and ordering by default}
\usage{
f_tbyl(data, vbl)
}
\description{
Provides a janitor distribution table but with formatting and ordering by default
}
