% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTools.R
\name{import_excel}
\alias{import_excel}
\title{Function to provide flexibility in import excel files}
\usage{
import_excel(path, password = "NONE", sheet = "NONE", skip = 0)
}
\description{
Function to provide flexibility in import excel files
}
