% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTools.R
\name{import_github}
\alias{import_github}
\title{Function to import github package with updated download setting}
\usage{
import_github(path)
}
\description{
Function to import github package with updated download setting
}
