% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomTools.R
\name{ViewExcel}
\alias{ViewExcel}
\title{Andrew Custom Package}
\usage{
ViewExcel(df = .Last.value, file = tempfile(fileext = ".csv"))
}
\description{
Provides a convenient function to view tables quickly in excel
}
